refactor: clean the code

+++ model/homepage.go
@@ -3,12 +3,12 @@ package model
 import "time"
 
 type ListPostsResponse struct {
-	ID           int       `json:"id"`
+	ID           int64     `json:"id"`
 	User         User      `json:"user"`
 	Content      string    `json:"content"`
 	ImageUrl     string    `json:"image_url"`
-	LikeCount    int       `json:"like_count"`
-	CommentCount int       `json:"comment_count"`
-	RepostCount  int       `json:"repost_count"`
+	LikeCount    int32     `json:"like_count"`
+	CommentCount int32     `json:"comment_count"`
+	RepostCount  int32     `json:"repost_count"`
 	UpdatedAt    time.Time `json:"updated_at"`
 }
